backend/requirements.txt
flask
pymongo
python-dotenv
flask-cors
pillow
qrcode[pil]